# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Películas & Series Latino",
    "version": "1.0.0",
    "description": "API REST en vivo para consultar catálogos estilo Netflix, metadatos y reproductores embeds verificados de películas, series, animes y doramas en audio Latino."
  },
  "servers": [
    {
      "url": "https://api-pelis-series-latino.vercel.app",
      "description": "Servidor de Producción Vercel"
    }
  ],
  "paths": {
...
    "/api/v1/search": {
      "get": {
        "summary": "Búsqueda general en tiempo real",
        "description": "Busca películas, series, animes y doramas por título o alias sobre el catálogo indexado, con coincidencia por SUBSTRING (ej. `ma` encuentra `Matilda`). El ranking prioriza los títulos que EMPIEZAN con el término. Respuesta LEAN por defecto (id, título, póster, año, rating, géneros) para máxima velocidad: los servidores, el cast y las temporadas se obtienen al abrir el detalle (`/api/v1/media/{id}` o `/api/v1/series/{id}`). Usa `full=true` para el objeto completo, o `compact=true` para la proyección compacta. `total_results` es el total REAL de coincidencias en el catálogo (no el acumulado de la página) y junto con `has_more` y `next_page` habilita el scroll infinito.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Término de búsqueda (ej. avatar, dragon ball, la casa del dragon)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Número de página, empezando en 1.",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Cantidad de resultados por página. Alias compatible: size. Máximo 100.",
            "schema": {
              "type": "integer",
              "default": 25,
              "maximum": 100
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Alias de limit para definir la cantidad de resultados por página. Máximo 100.",
            "schema": {
              "type": "integer",
              "default": 25,
              "maximum": 100
            }
          },
          {
            "name": "full",
            "in": "query",
            "description": "Si es true, devuelve el MediaItem completo (con cast, servers y seasons) en lugar de la proyección lean por defecto.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "compact",
            "in": "query",
            "description": "Si es true, devuelve la proyección compacta (incluye primary_stream y servers).",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Búsqueda exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/movies/search": {
      "get": {
        "summary": "Búsqueda general en tiempo real",
        "description": "Busca películas, series, animes y doramas por título o alias sobre el catálogo indexado, con coincidencia por SUBSTRING (ej. `ma` encuentra `Matilda`). El ranking prioriza los títulos que EMPIEZAN con el término. Respuesta LEAN por defecto (id, título, póster, año, rating, géneros) para máxima velocidad: los servidores, el cast y las temporadas se obtienen al abrir el detalle (`/api/v1/media/{id}` o `/api/v1/series/{id}`). Usa `full=true` para el objeto completo, o `compact=true` para la proyección compacta. `total_results` es el total REAL de coincidencias en el catálogo (no el acumulado de la página) y junto con `has_more` y `next_page` habilita el scroll infinito.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Término de búsqueda (ej. avatar, dragon ball, la casa del dragon)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Número de página, empezando en 1.",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Cantidad de resultados por página. Alias compatible: size. Máximo 100.",
            "schema": {
              "type": "integer",
              "default": 25,
              "maximum": 100
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Alias de limit para definir la cantidad de resultados por página. Máximo 100.",
            "schema": {
              "type": "integer",
              "default": 25,
              "maximum": 100
            }
          },
          {
            "name": "full",
            "in": "query",
            "description": "Si es true, devuelve el MediaItem completo (con cast, servers y seasons) en lugar de la proyección lean por defecto.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "compact",
            "in": "query",
            "description": "Si es true, devuelve la proyección compacta (incluye primary_stream y servers).",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Búsqueda exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ServerOption": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "quality": {
            "type": "string",
            "example": "1080p"
          },
          "language": {
            "type": "string",
            "example": "latino"
          },
          "embed_url": {
            "type": "string",
            "example": "https://vidhideplus.com/v/f1kgebbi5bly"
          },
          "status": {
            "type": "string",
            "enum": [
              "online",
              "offline"
            ],
            "description": "Estado de salud del servidor verificado en la capa de aplicación (Soft Errors). Los reproductores o clientes DEBEN IGNORAR O FILTRAR los servidores con status: 'offline'."
          },
          "last_checked": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Episode": {
        "type": "object",
        "properties": {
          "episode_number": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Los herederos del Dragón"
          },
          "still_path": {
            "type": "string",
            "nullable": true
          },
          "servers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServerOption"
            }
          }
        }
      },
      "Season": {
        "type": "object",
        "properties": {
          "season_number": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Temporada 1"
          },
          "episodes_count": {
            "type": "integer",
            "example": 10
          },
          "episodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Episode"
            }
          }
        }
      },
      "MediaItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "la-casa-del-dragon"
          },
          "tmdb_id": {
            "type": "integer"
          },
          "type": {
            "type": "string",
            "enum": [
              "movie",
              "tvseries"
            ],
            "description": "Determina si el contenido es película o serie. Para 'movie', consultar `servers`. Para 'tvseries', consultar `seasons` o la ruta de episodios."
          },
          "title": {
            "type": "string"
          },
          "original_title": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "release_date": {
            "type": "string"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "poster": {
            "type": "string",
            "nullable": true
          },
          "backdrop": {
            "type": "string",
            "nullable": true
          },
          "tagline": {
            "type": "string",
            "description": "Lema o eslogan oficial de TMDB"
          },
          "logo": {
            "type": "string",
            "nullable": true,
            "description": "Arte tipográfico del título (logo transparente de TMDB, es/en). Pensado para portadas y héroes."
          },
          "runtime": {
            "type": "integer",
            "nullable": true,
            "description": "Duración en minutos. En series, duración media del episodio."
          },
          "content_rating": {
            "type": "string",
            "nullable": true,
            "description": "Clasificación por edades oficial, priorizando mercados hispanohablantes (MX, AR, CL, ES, CO) y con US como respaldo."
          },
          "director": {
            "type": "string",
            "nullable": true,
            "description": "Director de la película."
          },
          "created_by": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            },
            "description": "Creadores de la serie."
          },
          "streams": {
            "$ref": "#/components/schemas/StreamsStatus"
          },
          "trailer": {
            "type": "string",
            "nullable": true,
            "description": "URL oficial de tráiler en YouTube (prioriza audio Latino)"
          },
          "cast": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Lista de nombres simples de actores del reparto (evita [object Object] en clientes)"
          },
          "cast_details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "character": {
                  "type": "string"
                },
                "photo": {
                  "type": "string",
                  "nullable": true
                }
              }
            },
            "description": "Lista detallada del reparto enriquecida desde TMDB con fotos y nombres de personaje"
          },
          "servers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServerOption"
            },
            "description": "Presente únicamente cuando type = movie"
          },
          "total_seasons": {
            "type": "integer",
            "description": "Presente únicamente cuando type = tvseries"
          },
          "total_episodes": {
            "type": "integer",
            "description": "Presente únicamente cuando type = tvseries"
          },
          "seasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Season"
            },
            "description": "Lista de temporadas y episodios. Presente únicamente cuando type = tvseries"
          }
        }
      },
      "StreamsStatus": {
        "type": "object",
        "description": "Estado de los enlaces de un título. La metadata se devuelve siempre al instante; cuando status es 'pending' los servidores todavía no están resueltos y el cliente debe pedirlos a `url` (normalmente al pulsar Reproducir).",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "pending"
            ]
          },
          "url": {
            "type": "string",
            "example": "/api/v1/media/scary-movie-6/streams"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Momento de la última resolución de enlaces. Por debajo de 24 h se sirven directamente de la base de datos."
          }
        }
      },
...
      "SearchResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "query": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "description": "Cantidad de elementos devueltos en esta página."
          },
          "results": {
            "type": "array",
            "description": "Por defecto, proyección LEAN de cada título (id, tmdb_id, type, title, original_title, poster, backdrop, release_date, rating, genres, subcategories) — SIN cast, servers ni seasons. Con full=true cada elemento es el MediaItem completo referenciado aquí.",
            "items": {
              "$ref": "#/components/schemas/MediaItem"
            }
          },
          "page": {
            "type": "integer",
            "description": "Número de página solicitado."
          },
          "limit": {
            "type": "integer",
            "description": "Cantidad máxima de resultados por página."
          },
          "total_results": {
            "type": "integer",
            "description": "Total REAL de coincidencias en el catálogo indexado para la consulta (substring), antes de paginar. Habilita el scroll infinito; no es el conteo acumulado de la página actual."
          },
          "has_more": {
            "type": "boolean",
            "description": "Indica si existen más resultados después de la página actual."
          },
          "next_page": {
            "type": "integer",
            "nullable": true,
            "description": "Número de la siguiente página cuando has_more es true; null si es la última."
          }
        }
      },
...
    }
  }
}